Move compute-related stuff out of util.py.
