+++ ocean_provider/routes/compute.py
@@ -14,15 +14,17 @@ from flask_sieve import validate
 from ocean_provider.requests_session import get_requests_session
 from ocean_provider.user_nonce import update_nonce
 from ocean_provider.utils.basics import LocalFileAdapter, get_provider_wallet, get_web3
+from ocean_provider.utils.compute import (
+    get_compute_endpoint,
+    get_compute_result_endpoint,
+    process_compute_request,
+    sign_for_compute,
+)
 from ocean_provider.utils.compute_environments import get_c2d_environments
 from ocean_provider.utils.error_responses import error_response
 from ocean_provider.utils.util import (
     build_download_response,
-    get_compute_endpoint,
-    get_compute_result_endpoint,
     get_request_data,
-    process_compute_request,
-    sign_for_compute,
 )
 from ocean_provider.validation.algo import WorkflowValidator
 from ocean_provider.validation.provider_requests import (
