+++ ocean_provider/run.py
@@ -15,7 +15,7 @@ from ocean_provider.utils.error_responses import strip_and_replace_urls
 from ocean_provider.myapp import app
 from ocean_provider.routes import services
 from ocean_provider.utils.basics import get_provider_wallet, get_web3
-from ocean_provider.utils.util import get_compute_info, get_request_data
+from ocean_provider.utils.util import get_request_data
 
 config = Config(filename=app.config["PROVIDER_CONFIG_FILE"])
 provider_url = config.get(ConfigSections.RESOURCES, "ocean_provider.url")
